move voice_components into question_answer folder
